# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "DeleteAccountRequest" : {
        "properties" : {
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LoginResponse" : {
        "properties" : {
          "principal" : {
            "$ref" : "#/components/schemas/UserPrincipal"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "principal", "token" ],
        "type" : "object"
      },
      "OtpRequest" : {
        "properties" : {
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "OtpVerifyRequest" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "email" ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TokenRefreshRequest" : {
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "token" ],
        "type" : "object"
      },
      "UpdateProfileRequest" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          }
        },
        "required" : [ "displayName" ],
        "type" : "object"
      },
      "UserPrincipal" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "role" : {
            "enum" : [ "GUEST", "USER", "ADMIN", "SUPER_ADMIN" ],
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "displayName", "id", "role", "username" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Content hub and knowledge management for the JVM ecosystem",
    "title" : "bytecode.news API",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/users/{username}/export" : {
      "get" : {
        "operationId" : "exportUserData_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "username",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "User data export"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Export a user's data for admin review",
        "tags" : [ "Admin - User Management" ]
      }
    },
...
    "/auth/account" : {
      "delete" : {
        "operationId" : "deleteAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Erase the authenticated user's account",
        "tags" : [ "Authentication" ]
      }
    },
    "/auth/export" : {
      "get" : {
        "operationId" : "exportUserData",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "User data export"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Export the authenticated user's data",
        "tags" : [ "Authentication" ]
      }
    },
...
    "/auth/otp/request" : {
      "post" : {
        "operationId" : "requestOtp",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Code sent if email is valid"
          }
        },
        "summary" : "Request a one-time sign-in code",
        "tags" : [ "Authentication" ]
      }
    },
    "/auth/otp/verify" : {
      "post" : {
        "operationId" : "verifyOtp",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpVerifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            },
            "description" : "Authentication successful"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Invalid or expired code"
          }
        },
        "summary" : "Verify a one-time sign-in code",
        "tags" : [ "Authentication" ]
      }
    },
    "/auth/profile" : {
      "put" : {
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Profile updated"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Update the authenticated user's profile",
        "tags" : [ "Authentication" ]
      }
    },
    "/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TokenRefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                }
              }
            },
            "description" : "Token refreshed"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Invalid or expired refresh token"
          }
        },
        "summary" : "Refresh an expired JWT token",
        "tags" : [ "Authentication" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:8080"
  } ]
}