# service-blog/src/main/kotlin/com/enigmastation/streampack/blog/model/UpdateProfileRequest.kt
/* Joseph B. Ottinger (C)2026 */
package com.enigmastation.streampack.blog.model

/** Request to update the authenticated user's display name */
data class UpdateProfileRequest(val displayName: String)